+++ plugin/core/protocol.py
@@ -6519,16 +6519,3 @@ CodeLensExtended = TypedDict('CodeLensExtended', {
 
 # Temporary for backward compatibility with LSP packages.
 RangeLsp = Range
-
-
-def kind_contains_other_kind(kind: str, other_kind: str) -> bool:
-    """
-    Check if `other_kind` is a sub-kind of `kind`.
-
-    The kind `"refactor.extract"` for example contains `"refactor.extract"` and ``"refactor.extract.function"`,
-    but not `"unicorn.refactor.extract"`, or `"refactor.extractAll"` or `refactor`.
-    """
-    if kind == other_kind:
-        return True
-    kind_len = len(kind)
-    return len(other_kind) > kind_len and other_kind.startswith(kind + '.')
